Introduce scroll and fade-in animations
Here we install react-reveal npm package and use the Fade component on the home, about, signup and login pages. Page headers will fade in on page load and as the user scrolls, other components will roll into view from the bottom of the window.

The scroll to top function in the footer was updated using an actual scroll animation.

This also updates some designs, such as adding a grey block behind the header, adjusting spacing and layout of the home & about pages, and other small changes.