Introduce state-controlled land ack modal
First pass at implementing a modal on the first page that starts open, stays open for 5s, and contains land acknowledgement information. TBD what this information will be. This modal is controlled by state, although the state is not persisted.

This PR accomplishes the set up of the modal - currently designed the same way as the team modals - and making it non-closeable for 5 seconds after page load. Currently the modal shows every time we navigte to the home page, despite being controlled by state.

Importantly, this starts a state slice for th broader user object.