Introduce redux-persist persisted store
Install and implement redux-persist to persist state to redux store
across page transitions and reloads. Storage is currently set to session
so that state will be reloaded on tab or window close, but not on
refresh or navigation. This is to better control when to show the land
ack modal and will also be useful for user authentication.